Modularize Bruh Beagle 2. This also finally makes the last counter dependent on the DOM.
