+++ index.html
@@ -57,7 +57,7 @@
     <img src="./bark.png" width="0" height="0" style="display: none;">
     <img src="./bagel.jpg" width="0" height="0" style="display: none;">
     <!-- Link to the code -->
-    <script src="./main.js" defer></script>
+    <script src="./main.mjs" type="module" defer></script>
   </body>
 
 </html>
